+++ test-suite-kotlin/src/test/kotlin/io/micronaut/docs/ioc/beans/Business.kt
@@ -1,4 +1,4 @@
-package io.micronaut.inject.visitor.beans
+package io.micronaut.docs.ioc.beans
 
 // tag::class[]
 import io.micronaut.core.annotation.Creator
