+++ test-suite-kotlin/src/test/kotlin/io/micronaut/docs/ioc/beans/Vehicle.kt
@@ -1,4 +1,4 @@
-package io.micronaut.inject.visitor.beans
+package io.micronaut.docs.ioc.beans
 
 // tag::class[]
 import io.micronaut.core.annotation.Creator
