+++ test-suite-kotlin/src/test/kotlin/io/micronaut/inject/visitor/beans/BeanIntrospectorSpec.kt
@@ -1,8 +1,6 @@
 package io.micronaut.inject.visitor.beans
 
-import io.micronaut.core.beans.BeanIntrospection
 import io.micronaut.core.beans.BeanIntrospector
-import junit.framework.TestCase
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.Test
 
@@ -36,19 +34,5 @@ class BeanIntrospectorSpec {
         assertEquals("newvalue", testBean.stuff)
     }
 
-    @Test
-    fun testVehicle() {
-        val introspection = BeanIntrospection.getIntrospection(Vehicle::class.java)
-        val vehicle = introspection.instantiate("Subaru", "WRX", 2)
-        TestCase.assertEquals("Subaru", vehicle.make)
-        TestCase.assertEquals("WRX", vehicle.model)
-        TestCase.assertEquals(2, vehicle.axels)
-    }
 
-    @Test
-    fun testBusiness() {
-        val introspection = BeanIntrospection.getIntrospection(Business::class.java)
-        val business = introspection.instantiate("Apple")
-        TestCase.assertEquals("Apple", business.name)
-    }
 }
